Cache everything
I had originally opted to *not* cache everything because I thought there
would be issues with faulty values clogging up the cache. I realized
this isn't a problem, faulty values will cause an exception and likely
end the test. The only way for faulty values to build up in the cache is
if the user invokes error and then squashes them, repeatedly. This is
very unlikely, they would have to do it deliberately.